agent/requirements.txt
langchain-google-genai>=4.2.0,<5.0.0
langchain-core>=1.2.7,<2.0.0
google-genai>=1.56.0,<2.0.0
pydantic>=2.7.0,<3.0.0
tavily-python==0.7.19
yfinance>=0.2.40,<2.0.0
requests>=2.32.0,<3.0.0
